Downgrade Coveralls plugin version
Currently the Travis job fails (https://travis-ci.org/alexandru/shade/jobs/97903369#L505)
to send coverage to Coveralls.io because of a bug in the plugin as mentioned in this
issue https://github.com/scoverage/sbt-coveralls/issues/73.

Let's try downgrading by 1 version to see if it gets fixed.

Downgrade Scoverage too as apparently that needs to be compatible
with our version of Coveralls

Add a note regarding SBT-Scoverage and SBT-coveralls compatibility