Switch to Trusty environment for Travis CI
The default Travis build environment is based on Ubuntu 12.04 (Precise
Pangolin), which includes memcached 1.4.13.  This version has several
bugs, including one where the incorrect default value is set when
incrementing a non-existent key.

This change moves the Travis environment to Google Compute Engine,
currently in beta, in order to run memcached inside a Docker container
based on the latest official memcached image, currently 1.4.25 as of this
commit.

We also change from openjdk7 to oraclejdk7 due to
travis-ci/travis-ci#5227, since the suggested workaround is also
non-functional currently due to travis-ci/travis-ci#5669.