Add ability to configure timeoutThreshold
By default, the timeout threshold for Spymemcached is set to 998, which means
that each connection will wait for 998 timeouts before considering the connection
to be dead and then re-establishing a new one.

This changeset adds the ability to configure the threshold from Shade.

Also, bump Spymemcached version to the latest one.